+++ lib/mx_merchant.dart
@@ -11,5 +11,6 @@ export 'src/models/payment/payment_response_model.dart';
 export 'src/models/payment/pos_data_model.dart';
 export 'src/models/payment/puchase_model.dart';
 export 'src/models/payment/risk_model.dart';
+export 'src/models/terminal/terminal_request_model.dart';
 export 'src/mx_merchant.dart';
-export 'src/utils/mxmerchant_exception.dart';
+export 'src/utils/mx_merchant_exception.dart';