+++ lib/src/models/terminal/terminal_payment_method_model.dart
@@ -1,0 +1,13 @@
+class MxTerminalPaymentMethodModel {
+  /// Credit
+  final bool credit;
+
+  /// debit
+  final bool debit;
+
+  MxTerminalPaymentMethodModel({this.credit = true, this.debit = true});
+
+  Map<String, dynamic> toJson() {
+    return {'credit': credit, 'debit': debit};
+  }
+}