+++ lib/src/utils/handle_response.dart
@@ -1,17 +1,15 @@
-import 'dart:convert';
-
 import 'package:dio/dio.dart';
 
-import 'mxmerchant_exception.dart';
+import 'mx_merchant_exception.dart';
 
-Map<String, dynamic> handleResponse(Response response) {
-  final body = response.data == null || (response.data != null && response.data.isEmpty)
-      ? Map<String, dynamic>.from({})
-      : Map<String, dynamic>.from(jsonDecode(response.data));
-
+T handleResponse<T>(Response response) {
   if ((response.statusCode ?? 500) >= 200 && (response.statusCode ?? 500) < 300) {
-    return body;
+    return response.data;
   }
 
-  throw MxMerchantException(statusCode: response.statusCode ?? 500, message: body['message']?.toString() ?? 'MX Merchant API error', response: body);
+  throw MxMerchantException(
+    statusCode: response.statusCode ?? 500,
+    message: response.data['message']?.toString() ?? 'MX Merchant API error',
+    response: response.data,
+  );
 }