+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.1-beta.3
+
+* Supported multiple base URLs of MX Merchant API.
+* Supported multiple authentication types of MX Merchant API (Basic and JWT).
+* Added terminal device service.
+
 ## 0.0.1-beta.1
 
 * Initial release.